requirements.txt
numpy==1.26.4
pandas==2.3.0
matplotlib==3.8.4
joblib==1.5.1
opencv-python-headless==4.11.0.86
Pillow==11.2.1
# scikit-learn==1.0.2
scikit-learn==1.3.2
scikit-image==0.25.2
ultralytics==8.3.173
PyYAML==6.0.2
imgaug==0.4.0
# torch==2.6.0+cu124
# torch==2.6.0
torch==2.3.0
# torchvision==0.21.0+cu124
torchvision==0.21.0
transformers==4.53.0
datasets==2.18.0
# datasets==3.6.0
# tf-keras==2.18.0
evaluate==0.4.5
streamlit==1.45.1